Update FunctionRpc.proto
Updated name to int_x and double_x since int and double are keywords in java and generated grpc code throws error if key words are used as variable names. 
Since GRPC relies on the number for comparing, it is better to rename the variable to anything but int and double here.